refactor(detail-article): refactor intent in article, refactor directory, refactor name
