+++ app/src/main/java/com/example/androidgroup4/ui/article/ArticleListFragment.kt
@@ -1,8 +1,7 @@
-package com.example.androidgroup4.ui.home
+package com.example.androidgroup4.ui.article
 
 import android.annotation.SuppressLint
 import android.content.Context
-import android.os.Bundle
 import android.view.*
 import android.view.inputmethod.InputMethodManager
 import android.widget.ImageView
@@ -12,25 +11,35 @@ import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.viewbinding.ViewBinding
 import com.bumptech.glide.Glide
-import com.example.androidgroup4.data.model.Article
-import com.example.androidgroup4.ui.MainActivity
 import com.example.androidgroup4.R
 import com.example.androidgroup4.base.BaseFragment
+import com.example.androidgroup4.data.model.Article
 import com.example.androidgroup4.databinding.FragmentHomeBinding
+import com.example.androidgroup4.ui.main.MainActivity
 import com.example.androidgroup4.ui.adapter.ArticleAdapter
 import com.example.androidgroup4.ui.auth.LoginActivity
-import com.example.androidgroup4.ui.detail.DetailArticleActivity
 import com.example.androidgroup4.utils.emptyString
 import com.synnapps.carouselview.ImageListener
 
-class HomeFragment: BaseFragment<FragmentHomeBinding>() {
+class ArticleListFragment : BaseFragment<FragmentHomeBinding>() {
 
-    var imageArray:ArrayList<Int> = ArrayList()
+    private var imageArray: ArrayList<Int> = ArrayList()
+
+    private var imageListener = ImageListener { position, imageView ->
+        imageView.scaleType = ImageView.ScaleType.CENTER_INSIDE
+        Glide.with(requireContext())
+            .load(imageArray[position])
+            .into(imageView)
+
+        imageView.setOnClickListener {
+            Toast.makeText(requireContext(), "Open Image", Toast.LENGTH_SHORT).show()   //tes
+        }
+    }
 
     private lateinit var articleAdapter: ArticleAdapter
 
-    override val bindingInflater: (LayoutInflater, ViewGroup?, Boolean) -> ViewBinding
-        = FragmentHomeBinding::inflate
+    override val bindingInflater: (LayoutInflater, ViewGroup?, Boolean) -> ViewBinding =
+        FragmentHomeBinding::inflate
 
     override fun initIntent() {
 
@@ -61,25 +70,10 @@ class HomeFragment: BaseFragment<FragmentHomeBinding>() {
         }
     }
 
-    var imageListener = ImageListener { position, imageView ->
-        imageView.scaleType = ImageView.ScaleType.CENTER_INSIDE
-        Glide.with(requireContext())
-            .load(imageArray[position])
-            .into(imageView)
-
-        imageView.setOnClickListener {
-            Toast.makeText(requireContext(), "Open Image", Toast.LENGTH_SHORT).show()   //tes
-        }
-    }
-
     @SuppressLint("ClickableViewAccessibility")
     override fun initAction() {
         articleAdapter.onItemClick = { article ->
-            // send data
-            val mBundle = Bundle()
-            mBundle.putParcelable(DetailArticleActivity.EXTRA_ARTICLE, article)
-            // untuk pindah fragmment
-            findNavController().navigate(R.id.action_navigation_home_to_detailArticleActivity, mBundle)
+            DetailArticleActivity.start(requireContext(), article)
         }
 
         with(binding) {
@@ -100,6 +94,7 @@ class HomeFragment: BaseFragment<FragmentHomeBinding>() {
 
 
     }
+
     fun View.hideKeyboard() {
         val imm = context.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
         imm.hideSoftInputFromWindow(windowToken, 0)
