+++ app/src/main/java/com/example/androidgroup4/ui/profile/ProfileFragment.kt
@@ -7,7 +7,7 @@ import com.example.androidgroup4.R
 import com.example.androidgroup4.base.BaseFragment
 import com.example.androidgroup4.data.model.Profile
 import com.example.androidgroup4.databinding.FragmentProfileBinding
-import com.example.androidgroup4.ui.MainActivity
+import com.example.androidgroup4.ui.main.MainActivity
 import com.example.androidgroup4.utils.constant.PreferenceKeys.IS_LOGIN
 import com.example.androidgroup4.utils.getAppPreferenceEditor
 import com.example.androidgroup4.utils.enum.GenderType
