+++ app/src/main/java/com/example/androidgroup4/utils/constant/BundleKeys.kt
@@ -15,4 +15,6 @@ object BundleKeys {
     const val DOCTOR = "DOCTOR"
 
     const val PROFILE = "PROFILE"
+
+    const val ARTICLE = "ARTICLE"
 }
